test: extract common metric logic
ref: #2318
