+++ visualization/app/codeCharta/ui/attributeSideBar/util/metricHelper.ts
@@ -1,0 +1,7 @@
+export type Metric = {
+	name: string
+	value: number
+	showAttributeTypeSelector: boolean
+}
+
+export const shouldShowAttributeTypeSelector = ({ isLeaf }: { isLeaf: boolean }) => !isLeaf
